[#6] - correcting test
